+++ changelog.md
@@ -1,1 +1,1 @@
-- Optimize energy pipes
+- Fixed energy pipe distribution